implement the generate method, using a trival integer partition alg.
The generate method take a Int (the length of grid) and a Sequence,
it will generate all possible solution matching this.

The intpartition will solve all [x_0..x_m] s.t.

  x_0+x_1+...+x_m = n where x_1...x_(m-1) >0 and x_0,x_m >= 0

it use a time-consuming algorithm and should be improved.

The original buildseq method is substituted with a new one.