add a QR Code Test Case No.14
The information contained in the QR Code is 'Picross',
but the recovery is not excat, they are also multi-solutions
like Test Case No.13 above.

Although the number of unknown discs is large, the uncertainty
of the whole graph may be low, there are some structure in it.

Maybe we can extract the information from the Sequence number?